deleted unused components and files
