--- src/components/ConfirmedGameDetail.js
@@ -1,9 +1,0 @@
-import { Box, Heading } from "@chakra-ui/react"
-
-export default function ConfirmedGameDetail() {
-    return (
-        <Box>
-            <Heading>ConfirmedGameDetail component</Heading>
-        </Box>
-    )
-}
